+++ photon-client/src/components/cameras/CameraCalibrationCard.vue
